+++ CHANGES.md
@@ -19,6 +19,7 @@ Changes in 12.8
 - Sanitize redirect_to: remove &delete_ok=1 in index.php
 - Add $html parameter & pass $pdf object by reference to the functions/PDF.php|pdf_stop_pdf action hook in PDF.php & Actions.php
 - CSS FlatSIS theme style checkbox & radio inputs in checkbox-radio.css & Gruntfile.js
+- List Courses to Secondary Teacher in StudentGrades.php
 
 Changes in 12.7.4
 -----------------